Add unit tests for auth and database functions
Add unit tests for the auth functions

This also led to some small bug fixes in other files, and the addition
of the chai-as-promised library (lets us test promises)

Add unit tests for database functions.

Export an sql function that executes arbritary sql

This is only intended to be used for testing purposes and other one offs.

Fix disabling compilation of tests

I meant to disable _linting_ for them, oops!

Clean up tests and fix invisible timeout bug

It the unit tests seem to be failing for no reason, there might be
a un-done() beforeAll of afterAll()

Add dome stub tests for database and oter utils

Complete database test spec