Swagger Update (#8)
* change some 'timepix'y strings to 'matchmyroute'

This should fix the wrong url showing up on the live system.

* Make swagger spec fully compliant.

Make swagger spec fully compliant.

Fix unneeded swagger description

* Made microservice framework respect security defs

This means that if you declare a security definition in the swagger spec, it will come up in params and you can do stuff with it. Perhaps in the future we could look at making it so that you add not just the definiton but the function too.

* Update gulp build to inject env vars into app.yaml

This means that when the live system is started on googel app engine,
the correct environment variables will be present.

* Remove commented out JWT parameter

We don't need these parameters any more, since the middleware
will give us the auth info.
This one was commented out while I was working on it, and I
forgot to delete it.