Add additional property and checks
Added days, which is a list of string days of the week

Also added checks for all the property types so we dont get invalid
owners or coordinates or anything like that. If it constructs without
throwing an error, it should be valid!

Update route data model to use geography

Also add the days parameter along with functions to convert it to/from
it's two formats:
 - An array of "monday", "tuesday", etc.
 - An integer with the n-th bit meaning the rider is available on the
 n-th day.

Added tests for new Route model

Add checks for RouteDataModel construction

This will reduce errors like Routes being created with no owners,
which cause problems down the line.

Add checks for property types of RouteDataModels

Add unit tests for the improved RouteDataModel