Add /buddyRequest [POST] endpoint
This endpoint lets us update the start/end location, time and names.
It does not let us update the status or reason - to do that one
must make a POST to /buddyRequest/status.
The reason for this separation is because:
 A) Changing status is actually fairly complex
 B) Both users can change the status (in some cases)
While we could put logic into this endpoint to accomplish A and B,
it would make the code rather complex and make bugs more likely
to appear.