Fix typos and create central getBuddyRequests
the Database.getBuddyRequests returns all of a user's sent/received
buddy requests, and the Database.getSentBuddyRequests +
Database.getReceivedBuddyRequests are now wrappers of this that
run a simple filter statement.

This means we only hit the database once, though if we find that
users have lots of buddy requests we may one day want to split
the sent/received functions back into full functions, for memory
efficiency.